+++ modules/spring-core/src/test/kotlin/com/esprito/spring/core/completion/properties/kotlin/ProjectSpringPropertiesCompletionContributorTest.kt
@@ -14,6 +14,8 @@ class ProjectSpringPropertiesCompletionContributorTest : AbstractSpringPropertie
         doTest(
             "mail.ert",
             "mail.external-settings",
+            "mail.external-settings.yet-property",
+            "mail.external-settings.camel-case-long-property-very-long-property",
             "mail.form-and-port",
             "mail.from",
             "mail.host-name",
@@ -37,6 +39,8 @@ class ProjectSpringPropertiesCompletionContributorTest : AbstractSpringPropertie
         doTest(
             "mail.ert",
             "mail.external-settings",
+            "mail.external-settings.yet-property",
+            "mail.external-settings.camel-case-long-property-very-long-property",
             "mail.form-and-port",
             "mail.host-name",
             "mail.nested-settings",
@@ -57,6 +61,8 @@ class ProjectSpringPropertiesCompletionContributorTest : AbstractSpringPropertie
             """mail.from=me
             mail.ert=ert
             mail.external-settings=ext
+            mail.external-settings.yet-property=yet
+            mail.external-settings.camel-case-long-property-very-long-property=camel
             mail.form-and-port=port
             mail.host-name=name
             mail.nested-settings=nested